requirements.txt
dataclasses==0.8
datasets==1.11.0
fire==0.4.0
huggingface-hub==0.0.8
nltk==3.6.6
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
psutil==5.9.1
pyahocorasick==1.4.4
pyarrow==6.0.1
pycparser==2.21
regex==2022.6.2
scikit-learn==0.24.2
scipy==1.5.4
seqeval==1.2.2
tokenizers==0.10.3
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
transformers==4.7.0

